refactoring architecture of the app, adding a repository
